event-nlp-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
scikit-learn
google-genai
requests
